my-bot/requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
comtypes==1.4.8
contourpy==1.2.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==5.1.1
docutils==0.21.2
Eel==0.16.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gensim==4.3.3
gevent==24.2.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
idna==3.6
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.6.1
keras==3.6.0
khmer-nltk==1.6
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.3
numpy==1.26.2
openai==1.57.2
opt-einsum==3.3.0
optree==0.12.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
pyarrow==18.0.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.5.2
pygbag==0.9.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
Pympler==1.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytweening==1.2.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
rich==13.7.1
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==69.2.0
six==1.16.0
sklearn-crfsuite==0.5.0
smart-open==7.0.5
sniffio==1.3.1
sockets==1.0.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-cpu==2.17.0
tensorflow_intel==2.18.0
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
vboxapi==1.0
wcwidth==0.2.12
Werkzeug==3.0.3
wheel==0.43.0
whichcraft==0.6.1
wrapt==1.16.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xxhash==3.5.0
yarl==1.17.2
you-get==0.4.1650
youtube-dl==2021.12.17
zope.event==5.0
zope.interface==6.2
